# swagger/swagger-app/dist/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "",
        "version": "1.0.0",
        "title": "Swagger Petstore",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "jing.zhan@ck-telecom.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "10.120.1.135:8080",
    "basePath": "/todo",
    "tags": [
        {
            "name": "user",
            "description": "Operations about user",
            "externalDocs": {
                "description": "Find out more about our store",
                "url": "http://swagger.io"
            }
        }
    ],
    "schemes": [
        "http"
    ],
    "paths": {
        "/user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "description": "This can only be done by the logged in user.",
                "operationId": "createUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "data",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "300": {
                        "description": "email was existed in database"
                    },
                    "400": {
                        "description": "pass data error!"
                    }
                }
            }
        },
        "/user/login": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Logs user into the system",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "The user name for login",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/UserInfo"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "type": "integer",
                                "format": "int32",
                                "description": "calls per hour allowed by the user"
                            },
                            "X-Expires-After": {
                                "type": "string",
                                "format": "date-time",
                                "description": "date in UTC when token expires"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Logs out current logged in user session",
                "description": "",
                "operationId": "logoutUser",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "确认登出用户",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "确认用户登录状态",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/{username}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get user by user name",
                "description": "",
                "operationId": "getUserByName",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The name that needs to be fetched. Use user1 for testing. ",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Updated user",
                "description": "This can only be done by the logged in user.",
                "operationId": "updateUser",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "name that need to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "data",
                        "description": "Updated user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    },
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteUser",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The name that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/project": {
            "post": {
                "tags": [
                    "project"
                ],
                "summary": "http://10.120.1.135:8082/todo/project",
                "description": "添加一个项目",
                "operationId": "addProject",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "info about new plan",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestProject"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "300": {
                        "description": "invalid user"
                    },
                    "400": {
                        "description": "exist project"
                    }
                }
            },
            "get": {
                "tags": [
                    "project"
                ],
                "summary": "http://10.120.1.135:8080/todo/project?email=jing.zhan@ck-telecom.com&token=aaaaa",
                "description": "获取用户参与的project",
                "operationId": "getprojects",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "用户email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token 确认登录状态",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "target_email",
                        "in": "query",
                        "type": "string",
                        "description": "要查询的用户",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success operation",
                        "schema": {
                            "$ref": "#/definitions/ResponseProject"
                        }
                    },
                    "300": {
                        "description": "invalid user"
                    },
                    "400": {
                        "description": "invaild parameters"
                    }
                }
            }
        },
        "/project/plan": {
            "get": {
                "summary": "http://10.120.1.135:8080/todo/project/plan?email=mozre&token=token&sprint=4&status=0",
                "description": "获取当前所有人的plan\n",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "确定用户",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "判断当前用户是否是处于active状态",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "sprint",
                        "in": "query",
                        "description": "指定获取的sprint",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "指定获取的状态的plan",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "plan"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ResponsePlan"
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "plan"
                ],
                "summary": "http://10.120.1.135:8080/todo/project/plan",
                "description": "添加一个计划",
                "operationId": "addplan",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "info about new plan",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestPlan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid form"
                    }
                }
            }
        },
        "/projcet/task": {
            "get": {
                "summary": "http://10.120.1.135:8080/todo/project/task?email=mozre&token=token",
                "description": "\"获取plan下的task\"\n",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "确定用户",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "判断当前用户是否是处于active状态",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "planid",
                        "in": "query",
                        "description": "获取指定plan的task",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "task"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ResponseTask"
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "task"
                ],
                "summary": "http://10.120.1.135:8080/todo/project/task",
                "description": "添加一个任务",
                "operationId": "add task",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "info about new task",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestTask"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid form"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "mem_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "mem_name": {
                    "type": "string"
                },
                "mem_email": {
                    "type": "string"
                },
                "mem_password": {
                    "type": "string"
                },
                "mem_phone": {
                    "type": "string"
                },
                "mem_icon": {
                    "type": "string"
                },
                "mem_level": {
                    "type": "integer"
                }
            },
            "xml": {
                "name": "User"
            }
        },
        "UserInfo": {
            "type": "object",
            "properties": {
                "mem_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "mem_name": {
                    "type": "string"
                },
                "mem_email": {
                    "type": "string"
                },
                "mem_password": {
                    "type": "string"
                },
                "mem_phone": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "UserInfo"
            }
        },
        "PlanSim": {
            "type": "object",
            "properties": {
                "plan_name": {
                    "type": "string"
                },
                "plan_decription": {
                    "type": "string"
                },
                "plan_deadline": {
                    "type": "string"
                },
                "plan_status": {
                    "type": "integer"
                }
            }
        },
        "ResponsePlan": {
            "type": "object",
            "properties": {
                "plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PlanSim"
                    }
                }
            }
        },
        "TaskSim": {
            "type": "object",
            "properties": {
                "task_title": {
                    "type": "string"
                },
                "task_description": {
                    "type": "string"
                },
                "task_priority": {
                    "type": "integer"
                },
                "task_status": {
                    "type": "integer"
                },
                "task_end_time": {
                    "type": "integer"
                }
            }
        },
        "ResponseTask": {
            "type": "object",
            "properties": {
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TaskSim"
                    }
                }
            }
        },
        "Task": {
            "type": "object",
            "properties": {
                "plan_id": {
                    "type": "string"
                },
                "men_id": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "string"
                },
                "task_title": {
                    "type": "string"
                },
                "task_content": {
                    "type": "string"
                },
                "task_type": {
                    "type": "integer"
                },
                "task_prioirty": {
                    "type": "integer"
                },
                "task_status": {
                    "type": "integer"
                },
                "task_start_time": {
                    "type": "string"
                },
                "task_pridict_time": {
                    "type": "string"
                },
                "task_remind_time": {
                    "type": "string"
                },
                "task_real_spend_time": {
                    "type": "string"
                }
            }
        },
        "RequestTask": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "task": {
                    "$ref": "#/definitions/Task"
                }
            }
        },
        "Plan": {
            "type": "object",
            "properties": {
                "plan_id": {
                    "type": "string"
                },
                "create_mem_id": {
                    "type": "integer"
                },
                "execute_mem_id": {
                    "type": "integer"
                },
                "project_id": {
                    "type": "string"
                },
                "plan_name": {
                    "type": "string"
                },
                "plan_description": {
                    "type": "string"
                },
                "plan_start_time": {
                    "type": "string"
                },
                "plan_end_time": {
                    "type": "string"
                },
                "plan_create_time": {
                    "type": "string"
                },
                "plan_state": {
                    "type": "integer",
                    "format": "int32"
                },
                "plan_real_spend_time": {
                    "type": "string"
                },
                "sprint": {
                    "type": "integer",
                    "format": "int32"
                },
                "plan_acomplish_progress": {
                    "type": "string"
                }
            }
        },
        "RequestPlan": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/Plan"
                }
            }
        },
        "Project": {
            "type": "object",
            "properties": {
                "project_id": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                },
                "projcet_description": {
                    "type": "string"
                },
                "mem_id": {
                    "type": "string"
                },
                "project_visibility": {
                    "type": "integer"
                }
            }
        },
        "RequestProject": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "postproject": {
                    "$ref": "#/definitions/Project"
                }
            }
        },
        "ResponseProject": {
            "type": "object",
            "properties": {
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Project"
                    }
                }
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}